+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Switched to `color-eyre` for error handling in the CLI.
+- Improved full download progress bar.
+- `Some(ProgressBar::Full)` is now the default for the library.
 
 ## [v0.4.5](https://github.com/epwalsh/rust-cached-path/releases/tag/v0.4.5) - 2020-09-15